:zap: Refactoring + filter for passports
