+++ modules/routers/external_ops.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends
 from yaml import YAMLError
 
 from ..exceptions import ConnectionTimeoutException, IncorrectAddressException, ParserException, UnhandledException
-from ..security import get_current_user
+from ..dependencies.security import get_current_user
 from ..utils import load_yaml
 
 router = APIRouter(dependencies=[Depends(get_current_user)])
